+++ ixmp/_config.py
@@ -61,6 +61,17 @@ def _locate(filename=None):
 
 def _platform_default():
     """Default values for the `platform` setting on BaseValues."""
+    try:
+        import ixmp4.conf
+
+        # Use configured ixmp4 storage directory
+        ixmp4_databases = ixmp4.conf.settings.storage_directory.joinpath("databases")
+    except ImportError:
+        # ixmp4 not installed or importable; construct a likely value
+        from platformdirs import user_data_path
+
+        ixmp4_databases = user_data_path("ixmp4").joinpath("databases")
+
     return {
         "default": "local",
         "local": {
@@ -70,11 +81,7 @@ def _platform_default():
         },
         "ixmp4-local": {
             "class": "ixmp4",
-            "dsn": (
-                "sqlite:///"
-                f"{Path.home().joinpath('.local', 'share', 'ixmp4', 'databases')}"
-                "/ixmp4-local.sqlite3"
-            ),
+            "dsn": f"sqlite:///{ixmp4_databases.joinpath('ixmp4-local.sqlite3')}",
         },
     }
 
@@ -367,18 +374,16 @@ class Config:
             if info not in self.values["platform"]:
                 raise ValueError(f"Cannot set unknown {repr(info)} as default platform")
         else:
-            from ixmp.backend import BACKENDS
+            from ixmp.backend import get_class
 
             _args = list(args)
 
             try:
                 # Get the backend class
                 cls = _args.pop(0)
-                backend_class = BACKENDS[cls]
+                backend_class = get_class(cls)
             except IndexError:
                 raise ValueError("Must give at least 1 arg: backend class")
-            except KeyError:
-                raise ValueError(f"No backend named {repr(cls)}")
 
             # Use the backend class' method to handle the arguments
             info = backend_class.handle_config(_args, kwargs)
@@ -421,9 +426,8 @@ class Config:
             return name, copy(self.values["platform"][name])
         except KeyError:
             raise ValueError(
-                f"platform name {repr(name)} not among "
-                + repr(sorted(self.values["platform"].keys()))
-                + f"\nfrom {self.path}"
+                f"platform name {name!r} not among {sorted(self.values['platform'])!r}"
+                f"\nfrom {self.path}"
             ) from None
 
     def remove_platform(self, name: str):