+++ ixmp/testing/__init__.py
@@ -47,7 +47,7 @@ import pint
 import pytest
 from click.testing import CliRunner
 
-from ixmp import BACKENDS, Platform, Scenario, cli
+from ixmp import Platform, Scenario, cli
 from ixmp import config as ixmp_config
 
 from .data import (
@@ -88,9 +88,6 @@ __all__ = [
     "tmp_env",
 ]
 
-# Parametrize platforms for jdbc and ixmp4
-backends = list(BACKENDS.keys())
-
 # Provide a skip marker since ixmp4 is not published for Python 3.9
 min_ixmp4_version = pytest.mark.skipif(
     sys.version_info < (3, 10), reason="ixmp4 requires Python 3.10 or higher"
@@ -105,17 +102,44 @@ def pytest_addoption(parser: pytest.Parser) -> None:
     parser.addoption(
         "--ixmp-jvm-mem",
         action="store",
-        help="Heap space to allocated for the JDBCBackend/JVM.",
+        default=-1,
+        help="Memory limit, in MiB, for the Java Virtual Machine (JVM) started by the "
+        "ixmp JDBCBackend",
+    )
+    parser.addoption(
+        "--ixmp-resource-limit",
+        action="store",
+        default="DATA:-1",
+        help=(
+            "Limit a Python resource via the ixmp.testing.resource_limit fixture. Use "
+            "e.g. 'DATA:500' to set RLIMIT_DATA to 500 MiB."
+        ),
     )
     parser.addoption(
         "--ixmp-user-config",
         action="store_true",
-        help="Use the user's existing config/'local' platform.",
+        help="Use the user's existing ixmp config, including platforms.",
     )
 
 
 def pytest_sessionstart(session: pytest.Session) -> None:
-    """Unset any configuration read from the user's directory."""
+    """Unset configuration read from the user's home directory or ``IXMP_DATA.
+
+    If :mod:`ixmp.testing` is required as a pytest plugin, this hook will *always* run,
+    even if the :func:`tmp_env` fixture is not used.
+
+    - *Unless* :program:`pytest … --ixmp-user-config` is given:
+
+      - The user's existing configuration is discarded and replaced with defaults.
+      - The path to the "local" (also "default") platform, which by default is in the
+        user's home directory, is cleared. This setting **must** be repopulated with a
+        valid path if :py:`ixmp.Platform()`, :py:`ixmp.Platform("default")`, or
+        :py:`ixmp.Platform("local")` is to be used. One way to do this is by using
+        the :func:`tmp_env` fixture.
+
+    - :data:`.jdbc._GC_AGGRESSIVE` is set to :any:`False` to disable aggressive garbage
+      collection, which can be slow.
+    """
     from ixmp.backend import jdbc
 
     if not session.config.option.ixmp_user_config:
@@ -124,7 +148,6 @@ def pytest_sessionstart(session: pytest.Session) -> None:
         # tmp_env below.
         ixmp_config.values["platform"]["local"].pop("path")
 
-    # Disable slow, aggressive garbage collection
     jdbc._GC_AGGRESSIVE = False
 
 
@@ -139,13 +162,12 @@ def pytest_report_header(config, start_path) -> str:
 def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
     """Parametrize tests for the two backend options."""
     if "backend" in metafunc.fixturenames:
-        markers = [m.name for m in metafunc.definition.iter_markers()]
+        import ixmp.backend
 
-        requested_backends = [marker for marker in markers if marker in backends]
+        ba = ixmp.backend.available()
+        markers = set(m.name for m in metafunc.definition.iter_markers())
 
-        _backends = requested_backends if bool(requested_backends) else backends
-
-        metafunc.parametrize("backend", _backends, indirect=True)
+        metafunc.parametrize("backend", sorted(set(ba) & markers) or ba, indirect=True)
 
 
 # Session-scoped fixtures
@@ -206,22 +228,50 @@ def test_mp(
 def tmp_env(
     pytestconfig: pytest.Config, tmp_path_factory: pytest.TempPathFactory
 ) -> Generator[os._Environ[str], Any, None]:
-    """Return the os.environ dict with the IXMP_DATA variable set.
+    """Temporary environment for testing.
 
-    IXMP_DATA will point to a temporary directory that is unique to the test session.
-    ixmp configuration (i.e. the 'config.json' file) can be written and read in this
-    directory without modifying the current user's configuration.
+    In this environment:
+
+    - The environment variable ``IXMP_DATA`` points to a temporary directory that is
+      unique to the test session.
+    - *Unless* :program:`pytest … --ixmp-user-config` is given, the "local" (also
+      "default") platform path is set to a file within the ``IXMP_DATA`` directory.
+    - The :file:`config.json` file is saved in ``IXMP_DATA``.
+
+    For :mod:`ixmp.tests`, this fixture is automatically invoked for every test session.
+    In downstream packages that use :mod:`ixmp.testing`, this *may not* be the case.
+
+    Returns
+    -------
+    dict
+        A reference to :data:`os.environ` with the ``IXMP_DATA`` key set.
     """
     base_temp = tmp_path_factory.getbasetemp()
     os.environ["IXMP_DATA"] = str(base_temp)
 
     if not pytestconfig.option.ixmp_user_config:
-        # Set the path for the default/local platform in the test directory
+        # Clear user's config. This (harmlessly) duplicates pytest_sessionstart, above.
+        ixmp_config.clear()
+        # Replace an automatic reference to the user's home directory with path to a
+        # default/local platform within the pytest temporary directory
         localdb = base_temp.joinpath("localdb", "default")
         ixmp_config.values["platform"]["local"]["path"] = localdb
 
     # Save for other processes
     ixmp_config.save()
+
+    try:
+        import ixmp4.conf
+
+        # Replace an automatic reference to the user's home directory with a
+        # subdirectory of the pytest temporary directory
+        ixmp4.conf.settings.storage_directory = base_temp.joinpath("ixmp4")
+        # Ensure this directory and a further subdirectory "databases" exist
+        ixmp4.conf.settings.storage_directory.joinpath("databases").mkdir(
+            parents=True, exist_ok=True
+        )
+    except ImportError:
+        pass
 
     yield os.environ
 
@@ -454,19 +504,6 @@ def create_test_platform(tmp_path, data_path, name, **properties):
 # Private utilities
 
 
-# NOTE The test suite doesn't need all units/regions defined by default
-# Only those marked with 'tutorials' are needed for the MESSAGE tutorials
-def _setup_ixmp4_platform(mp: Platform) -> None:
-    """Set up an ixmp4-backed Platform with things hardcoded in Java."""
-    mp.add_unit("???", comment="As pre-defined in Java.")
-    mp.add_unit("GWa", comment="As pre-defined in Java.")
-    mp.add_unit("USD/kWa", comment="As pre-defined in Java.")
-    mp.add_unit("cases", comment="As pre-defined in Java.")  # tutorials
-    mp.add_unit("kg", comment="As pre-defined in Java.")
-    mp.add_unit("km", comment="As pre-defined in Java.")  # tutorials
-    mp.add_region(region="World", hierarchy="common")
-
-
 def _platform_fixture(
     request: pytest.FixtureRequest,
     tmp_env,
@@ -478,38 +515,40 @@ def _platform_fixture(
     # Remove '/' so that the name can be used in URL tests.
     platform_name = request.node.nodeid.replace("/", " ")
 
-    # Add a platform
+    # Construct positional and keyword arguments to Config.add_platform()
     if backend == "jdbc":
-        ixmp_config.add_platform(
-            platform_name, backend, "hsqldb", url=f"jdbc:hsqldb:mem:{platform_name}"
-        )
+        args = ["hsqldb"]
+        kwargs: dict[str, Any] = dict(url=f"jdbc:hsqldb:mem:{platform_name}")
     elif backend == "ixmp4":
-        import ixmp4.conf
-        from ixmp4.core.exceptions import PlatformNotUnique
+        args = []
+        kwargs = dict(
+            ixmp4_name=platform_name, dsn="sqlite:///:memory:", jdbc_compat=True
+        )
 
-        # Add to or get from ixmp4 an in-memory DB/Platform
-        dsn = "sqlite:///:memory:"
-        try:
-            ixmp4.conf.settings.toml.add_platform(name=platform_name, dsn=dsn)
-        except PlatformNotUnique:
-            pass
-
-        # Add ixmp4 backend to ixmp platforms
-        ixmp_config.add_platform(platform_name, backend, _name=platform_name)
+    # Add platform to ixmp configuration
+    ixmp_config.add_platform(platform_name, backend, *args, **kwargs)
 
     # Launch Platform
-    mp = Platform(name=platform_name, backend=backend)
-    if backend == "ixmp4":
-        _setup_ixmp4_platform(mp=mp)
+    mp = Platform(name=platform_name)
+
     yield mp
 
-    # Teardown: don't show log messages when destroying the platform, even if
-    # the test using the fixture modified the log level
+    # Teardown: don't show log messages when destroying the platform, even if the test
+    # using the fixture modified the log level
     mp._backend.set_log_level(logging.CRITICAL)
+
     del mp
 
-    # Remove from config
+    # Remove from configuration
     ixmp_config.remove_platform(platform_name)
 
     if backend == "ixmp4":
-        ixmp4.conf.settings.toml.remove_platform(key=platform_name)
+        import ixmp4.conf
+        from ixmp4.core.exceptions import PlatformNotFound
+
+        try:
+            ixmp4.conf.settings.toml.remove_platform(key=platform_name)
+        except PlatformNotFound:
+            # This occurs e.g. if `mp` was not used in a way that triggered addition of
+            # the name & DSN to the ixmp4 configuration.
+            pass